Small improvements in multiple files

+++ math/modular_arithmetic.cpp
@@ -2,7 +2,7 @@ const int M = 1e9 + 7;
 int add(int a, int b){ return a+b < M ? a+b : a+b-M; }
 int sub(int a, int b){ return a-b >= 0 ? a-b : a-b+M; }
 int mul(int a, int b){ return int(ll(a)*b % M); }
-int pot(int b, int e){ // O(log e)
+int pot(int b, ll e){ // O(log e)
 	int r=1;
     while (e) {
         if (e&1) r = mul(r,b);
